+++ R/ggpmisc.R
@@ -44,9 +44,7 @@
 #' \url{https://ggplot2.tidyverse.org/}\cr Package 'ggplot2' source code at
 #' \url{https://github.com/tidyverse/ggplot2}
 #'
-#' @import scales grid ggplot2
-#' @importFrom ggplot2 ggplot
-#' @importFrom magrittr %>%
+#' @import scales grid ggpp
 #' @importFrom rlang .data
 #'
 #' @note The signatures of \code{stat_peaks()} and \code{stat_valleys()} are